requirements.txt
streamlit
pandas
plotly
sqlalchemy
python-dateutil
psycopg2-binary
